+++ management/management.gen_test.go
@@ -2725,6 +2725,13 @@ func TestClient_GetTokenQuota(tt *testing.T) {
 	c.GetTokenQuota()
 }
 
+func TestClient_GetTokenVaultPrivilegedAccess(tt *testing.T) {
+	c := &Client{}
+	c.GetTokenVaultPrivilegedAccess()
+	c = nil
+	c.GetTokenVaultPrivilegedAccess()
+}
+
 func TestClient_GetWebOrigins(tt *testing.T) {
 	var zeroValue []string
 	c := &Client{WebOrigins: &zeroValue}
@@ -3475,6 +3482,72 @@ func TestClientTokenExchange_GetAllowAnyProfileOfType(tt *testing.T) {
 func TestClientTokenExchange_String(t *testing.T) {
 	var rawJSON json.RawMessage
 	v := &ClientTokenExchange{}
+	if err := json.Unmarshal([]byte(v.String()), &rawJSON); err != nil {
+		t.Errorf("failed to produce a valid json")
+	}
+}
+
+func TestClientTokenVaultPrivilegedAccess_GetCredentials(tt *testing.T) {
+	var zeroValue []Credential
+	c := &ClientTokenVaultPrivilegedAccess{Credentials: &zeroValue}
+	c.GetCredentials()
+	c = &ClientTokenVaultPrivilegedAccess{}
+	c.GetCredentials()
+	c = nil
+	c.GetCredentials()
+}
+
+func TestClientTokenVaultPrivilegedAccess_GetGrants(tt *testing.T) {
+	var zeroValue []ClientTokenVaultPrivilegedGrant
+	c := &ClientTokenVaultPrivilegedAccess{Grants: &zeroValue}
+	c.GetGrants()
+	c = &ClientTokenVaultPrivilegedAccess{}
+	c.GetGrants()
+	c = nil
+	c.GetGrants()
+}
+
+func TestClientTokenVaultPrivilegedAccess_GetIPAllowlist(tt *testing.T) {
+	var zeroValue []string
+	c := &ClientTokenVaultPrivilegedAccess{IPAllowlist: &zeroValue}
+	c.GetIPAllowlist()
+	c = &ClientTokenVaultPrivilegedAccess{}
+	c.GetIPAllowlist()
+	c = nil
+	c.GetIPAllowlist()
+}
+
+func TestClientTokenVaultPrivilegedAccess_String(t *testing.T) {
+	var rawJSON json.RawMessage
+	v := &ClientTokenVaultPrivilegedAccess{}
+	if err := json.Unmarshal([]byte(v.String()), &rawJSON); err != nil {
+		t.Errorf("failed to produce a valid json")
+	}
+}
+
+func TestClientTokenVaultPrivilegedGrant_GetConnection(tt *testing.T) {
+	var zeroValue string
+	c := &ClientTokenVaultPrivilegedGrant{Connection: &zeroValue}
+	c.GetConnection()
+	c = &ClientTokenVaultPrivilegedGrant{}
+	c.GetConnection()
+	c = nil
+	c.GetConnection()
+}
+
+func TestClientTokenVaultPrivilegedGrant_GetScopes(tt *testing.T) {
+	var zeroValue []string
+	c := &ClientTokenVaultPrivilegedGrant{Scopes: &zeroValue}
+	c.GetScopes()
+	c = &ClientTokenVaultPrivilegedGrant{}
+	c.GetScopes()
+	c = nil
+	c.GetScopes()
+}
+
+func TestClientTokenVaultPrivilegedGrant_String(t *testing.T) {
+	var rawJSON json.RawMessage
+	v := &ClientTokenVaultPrivilegedGrant{}
 	if err := json.Unmarshal([]byte(v.String()), &rawJSON); err != nil {
 		t.Errorf("failed to produce a valid json")
 	}